pyproject.toml
[project]
authors = [{name = "GJCav", email = "micojcav@outlook.com"}]
name = "tensor2frame-bench"
requires-python = "==3.13"
version = "0.1.0"

dependencies = [
    "pyvista",
    "pygfx",
    "glfw",
    "pytorch3d @ https://github.com/MiroPsota/torch_packages_builder/releases/download/pytorch3d-0.7.8/pytorch3d-0.7.8%2Bpt2.7.1cu118-cp313-cp313-linux_x86_64.whl",

    "numpy",
    "pylinalg",
    "scikit-image",
    "attrs",
    "alpenstock",
    "humanize",
    "roma", # handle rotations
    "imageio",
    "imageio-ffmpeg",
]


[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
tensor2frame_bench = { path = ".", editable = true }
torch = { version = "==2.7.1", index = "https://download.pytorch.org/whl/cu118" }
torchaudio = { version = "==2.7.1", index = "https://download.pytorch.org/whl/cu118" }
torchvision = { version = "==0.22.1", index = "https://download.pytorch.org/whl/cu118" }

[tool.pixi.tasks]

[tool.hatch.metadata]
allow-direct-references = true